adding manifests & updating app.py

+++ app.py
@@ -3,9 +3,12 @@
 import os
 import fasttext
 import json
+import logging
 from flask import Flask, request
 from dotenv import load_dotenv, find_dotenv
 import boto3
+
+_LOGGER = logging.getLogger("github-labeler")
 
 # load environment variables
 load_dotenv(find_dotenv())
@@ -25,7 +28,7 @@ if use_ceph:
         endpoint_url=s3_endpoint_url,
     )
 
-application = Flask(__name__)
+application = Flask("github-labeler")
 
 # read in bots
 
@@ -48,42 +51,46 @@ with open("labellist.txt", "r") as h:
     labels = [lbl.replace("\n", "") for lbl in h.readlines()]
 
 
-@application.route("/predict", methods=["POST"])
+@application.route("/predict", methods=["POST", "GET"])
 def predict():
     """Take json data of title, body, created_by and returns a tab separated list of strings."""
-    data = request.get_json() or "{}"
-    data = json.loads(data)
-    title, body, creator = data["title"], data["body"], data["created_by"]
-    if creator in bots:
-        return ""
-    ret = []
-    body = body.replace("\r", "<R>").replace("\n", "<N>")
-    input_ = title + "<SEP>" + body
+    if request.method == "POST":
+        data = request.get_json() or "{}"
+        data = json.loads(data)
+        title, body, creator = data["title"], data["body"], data["created_by"]
+        if creator in bots:
+            return ""
+        ret = []
+        body = body.replace("\r", "<R>").replace("\n", "<N>")
+        input_ = title + "<SEP>" + body
 
-    if use_ceph:
-        for lbl in labels:
-            path = os.path.join("saved_models", lbl.replace("/", "_") + ".bin")
-            model = s3.get_object(Bucket=s3_bucket, Key=path)
-            with open(path, "wb") as f:
-                for i in model["Body"]:
-                    f.write(i)
-            model = fasttext.load_model(path)
-            pred, prob = model.predict(input_)
-            if pred[0] == "__label__0" and prob > threshold:
-                print(prob)
-                ret.append(lbl)
-            os.remove(path)
+        if use_ceph:
+            for lbl in labels:
+                path = os.path.join("saved_models", lbl.replace("/", "_") + ".bin")
+                model = s3.get_object(Bucket=s3_bucket, Key=path)
+                with open(path, "wb") as f:
+                    for i in model["Body"]:
+                        f.write(i)
+                model = fasttext.load_model(path)
+                pred, prob = model.predict(input_)
+                if pred[0] == "__label__0" and prob > threshold:
+                    print(prob)
+                    ret.append(lbl)
+                os.remove(path)
 
+        else:
+            for lbl in labels:
+                path = os.path.join("saved_models", lbl.replace("/", "_") + ".bin")
+                model = fasttext.load_model(path)
+                pred, prob = model.predict(input_)
+                if pred[0] == "__label__0" and prob > threshold:
+                    print(prob)
+                    ret.append(lbl)
+        return "\t".join(ret)
     else:
-        for lbl in labels:
-            path = os.path.join("saved_models", lbl.replace("/", "_") + ".bin")
-            model = fasttext.load_model(path)
-            pred, prob = model.predict(input_)
-            if pred[0] == "__label__0" and prob > threshold:
-                print(prob)
-                ret.append(lbl)
-    return "\t".join(ret)
+        return "ALL GOOD"
 
 
 if __name__ == "__main__":
-    application.run(debug=True)
+    _LOGGER.debug("Debug mode is on")
+    application.run(host="0.0.0.0", port=8080)
